requirements.txt
streamlit
streamlit-image-select
streamlit-image-coordinates
opencv-python
pycocotools
matplotlib
onnxruntime
onnx
git+https://github.com/facebookresearch/segment-anything.git
graphviz
torch
torchvision
torchaudio
torchview
torchsummary
torchmetrics
pandas
tqdm
segmentation-models-pytorch
wandb
albumentations
numpy
pillow